Refactor part 1
