+++ lib/mixpanel-node.js
@@ -7,151 +7,51 @@
     Released under the MIT license.
 */
 
-var http            = require('http'),
-    https           = require('https'),
-    querystring     = require('querystring'),
-    Buffer          = require('buffer').Buffer,
-    util            = require('util'),
-    HttpsProxyAgent = require('https-proxy-agent');
-
-var async_all = require('./utils').async_all;
-
-var REQUEST_LIBS = {
-    http: http,
-    https: https
-};
-
-var create_proxy_agent = function() {
-    var proxyPath = process.env.HTTPS_PROXY || process.env.HTTP_PROXY;
-    return proxyPath ? new HttpsProxyAgent(proxyPath) : null;
-}
+const {async_all, ensure_timestamp, create_config, create_send_request_func} = require('./utils');
+const {merge_modifiers, _set, _delete_profile, _unset, _union, _remove} = require('./profile_helpers');
+const {create_group_funcs} = require('./groups');
 
 var create_client = function(token, config) {
-    var metrics = {};
-    var proxyAgent = create_proxy_agent();
-
     // mixpanel constants
-    var MAX_BATCH_SIZE = 50,
-        TRACK_AGE_LIMIT = 60 * 60 * 24 * 5;
+    const MAX_BATCH_SIZE = 50;
+    const TRACK_AGE_LIMIT = 60 * 60 * 24 * 5;
+    const ENGAGE_ENDPOINT = '/engage';
 
     if(!token) {
         throw new Error("The Mixpanel Client needs a Mixpanel token: `init(token)`");
     }
 
-    // Default config
-    metrics.config = {
-        test: false,
-        debug: false,
-        verbose: false,
-        host: 'api.mixpanel.com',
-        protocol: 'https',
-        path: '',
-    };
-
-    metrics.token = token;
+    let metrics = {token};
 
     /**
-     * sends an async GET or POST request to mixpanel
-     * for batch processes data must be send in the body of a POST
-     * @param {object} options
-     * @param {string} options.endpoint
-     * @param {object} options.data         the data to send in the request
-     * @param {string} [options.method]     e.g. `get` or `post`, defaults to `get`
-     * @param {function} callback           called on request completion or error
-     */
-    metrics.send_request = function(options, callback) {
-        callback = callback || function() {};
+        set_config(config)
+        ---
+        Modifies the mixpanel config
 
-        var content = Buffer.from(JSON.stringify(options.data)).toString('base64'),
-            endpoint = options.endpoint,
-            method = (options.method || 'GET').toUpperCase(),
-            query_params = {
-                'ip': 0,
-                'verbose': metrics.config.verbose ? 1 : 0
-            },
-            key = metrics.config.key,
-            request_lib = REQUEST_LIBS[metrics.config.protocol],
-            request_options = {
-                host: metrics.config.host,
-                port: metrics.config.port,
-                headers: {},
-                method: method
-            },
-            request;
+        config:object       an object with properties to override in the
+                            mixpanel client config
+    */
+    metrics.set_config = (new_config) => {
+        metrics.config = create_config(new_config);
+        /**
+         * sends an async GET or POST request to mixpanel
+         * for batch processes data must be send in the body of a POST
+         * @param {object} options
+         * @param {string} options.endpoint
+         * @param {object} options.data         the data to send in the request
+         * @param {string} [options.method]     e.g. `get` or `post`, defaults to `get`
+         * @param {function} callback           called on request completion or error
+         */
+        metrics.send_request = create_send_request_func(metrics.config);
 
-        if (!request_lib) {
-            throw new Error(
-                "Mixpanel Initialization Error: Unsupported protocol " + metrics.config.protocol + ". " +
-                "Supported protocols are: " + Object.keys(REQUEST_LIBS)
-            );
-        }
-
-
-        if (method === 'POST') {
-            content = 'data=' + content;
-            request_options.headers['Content-Type'] = 'application/x-www-form-urlencoded';
-            request_options.headers['Content-Length'] = Buffer.byteLength(content);
-        } else if (method === 'GET') {
-            query_params.data = content;
-        }
-
-
-        // add `key` query params
-        if (key) {
-            query_params.api_key = key;
-        } else if (endpoint === '/import') {
-            throw new Error("The Mixpanel Client needs a Mixpanel api key when importing old events: `init(token, { key: ... })`");
-        }
-
-        if (proxyAgent) {
-            request_options.agent = proxyAgent;
-        }
-
-        if (metrics.config.test) {
-            query_params.test = 1;
-        }
-
-        request_options.path = metrics.config.path + endpoint + "?" + querystring.stringify(query_params);
-
-        request = request_lib.request(request_options, function(res) {
-            var data = "";
-            res.on('data', function(chunk) {
-                data += chunk;
-            });
-
-            res.on('end', function() {
-                var e;
-                if (metrics.config.verbose) {
-                    try {
-                        var result = JSON.parse(data);
-                        if(result.status != 1) {
-                            e = new Error("Mixpanel Server Error: " + result.error);
-                        }
-                    }
-                    catch(ex) {
-                        e = new Error("Could not parse response from Mixpanel");
-                    }
-                }
-                else {
-                    e = (data !== '1') ? new Error("Mixpanel Server Error: " + data) : undefined;
-                }
-
-                callback(e);
-            });
+        metrics.groups = create_group_funcs({
+            token: metrics.token,
+            config: metrics.config,
+            send_request: metrics.send_request,
         });
-
-        request.on('error', function(e) {
-            if (metrics.config.debug) {
-                console.log("Got Error: " + e.message);
-            }
-            callback(e);
-        });
-
-        if (method === 'POST') {
-            request.write(content);
-        }
-        request.end();
     };
+    metrics.set_config(config);
+
 
     /**
      * Send an event to Mixpanel, using the specified endpoint (e.g., track/import)
@@ -174,18 +74,6 @@ var create_client = function(token, config) {
         }
 
         metrics.send_request({ method: "GET", endpoint: endpoint, data: data }, callback);
-    };
-
-    /**
-     * Validate type of time property, and convert to Unix timestamp if necessary
-     * @param {Date|number} time - value to check
-     * @returns {number} Unix timestamp
-     */
-    var ensure_timestamp = function(time) {
-        if (!(time instanceof Date || typeof time === "number")) {
-            throw new Error("`time` property must be a Date or Unix timestamp and is only required for `import` endpoint");
-        }
-        return time instanceof Date ? Math.floor(time.getTime() / 1000) : time;
     };
 
     /**
@@ -453,27 +341,8 @@ var create_client = function(token, config) {
 
         */
         set_once: function(distinct_id, prop, to, modifiers, callback) {
-            var $set = {};
-
-            if (typeof(prop) === 'object') {
-                if (typeof(to) === 'object') {
-                    callback = modifiers;
-                    modifiers = to;
-                } else {
-                    callback = to;
-                }
-                $set = prop;
-            } else {
-                $set[prop] = to;
-                if (typeof(modifiers) === 'function' || !modifiers) {
-                    callback = modifiers;
-                }
-            }
-
-            modifiers = modifiers || {};
-            modifiers.set_once = true;
-
-            this._set(distinct_id, $set, callback, modifiers);
+            const identifiers = {$distinct_id: distinct_id};
+            _set(prop, to, modifiers, callback, {identifiers, set_once: true, token: metrics.token, endpoint: ENGAGE_ENDPOINT, send_request: metrics.send_request, config: metrics.config});
         },
 
         /**
@@ -491,56 +360,10 @@ var create_client = function(token, config) {
                 });
         */
         set: function(distinct_id, prop, to, modifiers, callback) {
-            var $set = {};
-
-            if (typeof(prop) === 'object') {
-                if (typeof(to) === 'object') {
-                    callback = modifiers;
-                    modifiers = to;
-                } else {
-                    callback = to;
-                }
-                $set = prop;
-            } else {
-                $set[prop] = to;
-                if (typeof(modifiers) === 'function' || !modifiers) {
-                    callback = modifiers;
-                }
-            }
-
-            this._set(distinct_id, $set, callback, modifiers);
+            const identifiers = {$distinct_id: distinct_id};
+            _set(prop, to, modifiers, callback, {identifiers, token: metrics.token, endpoint: ENGAGE_ENDPOINT, send_request: metrics.send_request, config: metrics.config});
         },
 
-        // used internally by set and set_once
-        _set: function(distinct_id, $set, callback, options) {
-            options = options || {};
-            var set_key = (options && options.set_once) ? "$set_once" : "$set";
-
-            var data = {
-                '$token': metrics.token,
-                '$distinct_id': distinct_id
-            };
-            data[set_key] = $set;
-
-            if ('ip' in $set) {
-                data.$ip = $set.ip;
-                delete $set.ip;
-            }
-
-            if ($set.$ignore_time) {
-                data.$ignore_time = $set.$ignore_time;
-                delete $set.$ignore_time;
-            }
-
-            data = merge_modifiers(data, options);
-
-            if (metrics.config.debug) {
-                console.log("Sending the following data to Mixpanel (Engage):");
-                console.log(data);
-            }
-
-            metrics.send_request({ method: "GET", endpoint: "/engage", data: data }, callback);
-        },
 
         /**
             people.increment(distinct_id, prop, by, modifiers, callback)
@@ -774,21 +597,8 @@ var create_client = function(token, config) {
                 mixpanel.people.delete_user('bob');
         */
         delete_user: function(distinct_id, modifiers, callback) {
-            var data = {
-                '$delete': '',
-                '$token': metrics.token,
-                '$distinct_id': distinct_id
-            };
-
-            if (typeof(modifiers) === 'function') { callback = modifiers; }
-
-            data = merge_modifiers(data, modifiers);
-
-            if (metrics.config.debug) {
-                console.log("Deleting the user from engage:", distinct_id);
-            }
-
-            metrics.send_request({ method: "GET", endpoint: "/engage", data: data }, callback);
+            const identifiers = {$distinct_id: distinct_id};
+            _delete_profile({identifiers, modifiers, callback, endpoint: ENGAGE_ENDPOINT, token: metrics.token, send_request: metrics.send_request, config: metrics.config});
         },
 
         /**
@@ -803,51 +613,17 @@ var create_client = function(token, config) {
               mixpanel.people.remove('bob', {'browsers': 'chrome', 'os': 'linux'});
         */
         remove: function(distinct_id, data, modifiers, callback) {
-            var $remove = {};
-
-            if (typeof(data) !== 'object' || util.isArray(data)) {
-                if (metrics.config.debug) {
-                    console.error("Invalid value passed to mixpanel.people.remove - data must be an object with" +
-                      " scalar values");
-                }
-                return;
-            }
-
-            Object.keys(data).forEach(function(key) {
-                var val = data[key];
-                if (typeof(val) === 'string' || typeof(val) === 'number') {
-                    $remove[key] = val;
-                } else {
-                    if (metrics.config.debug) {
-                        console.error("Invalid argument passed to mixpanel.people.remove - values must be scalar");
-                        console.error("Passed " + key + ':', val);
-                    }
-                    return;
-                }
-            });
-
-            if (Object.keys($remove).length === 0) {
-                return;
-            }
-
-            data = {
-                '$remove': $remove,
-                '$token': metrics.token,
-                '$distinct_id': distinct_id
-            };
-
-            if (typeof(modifiers) === 'function') {
-                callback = modifiers;
-            }
-
-            data = merge_modifiers(data, modifiers);
-
-            if (metrics.config.debug) {
-                console.log("Sending the following data to Mixpanel (Engage):");
-                console.log(data);
-            }
-
-            metrics.send_request({ method: "GET", endpoint: "/engage", data: data }, callback);
+            const identifiers = {'$distinct_id': distinct_id};
+            _remove({
+                identifiers,
+                data,
+                modifiers,
+                callback,
+                endpoint: ENGAGE_ENDPOINT,
+                token: metrics.token,
+                send_request: metrics.send_request,
+                config: metrics.config,
+            })
         },
 
         /**
@@ -862,57 +638,8 @@ var create_client = function(token, config) {
             mixpanel.people.union('bob', {'browsers': ['chrome'], os: ['linux']});
          */
         union: function(distinct_id, data, modifiers, callback) {
-            var $union = {};
-
-            if (typeof(data) !== 'object' || util.isArray(data)) {
-                if (metrics.config.debug) {
-                    console.error("Invalid value passed to mixpanel.people.union - data must be an object with scalar or array values");
-                }
-                return;
-            }
-
-            Object.keys(data).forEach(function(key) {
-                var val = data[key];
-                if (util.isArray(val)) {
-                    var merge_values = val.filter(function(v) {
-                        return typeof(v) === 'string' || typeof(v) === 'number';
-                    });
-                    if (merge_values.length > 0) {
-                        $union[key] = merge_values;
-                    }
-                } else if (typeof(val) === 'string' || typeof(val) === 'number') {
-                    $union[key] = [val];
-                } else {
-                    if (metrics.config.debug) {
-                        console.error("Invalid argument passed to mixpanel.people.union - values must be a scalar value or array");
-                        console.error("Passed " + key + ':', val);
-                    }
-                    return;
-                }
-            });
-
-            if (Object.keys($union).length === 0) {
-                return;
-            }
-
-            data = {
-                '$union': $union,
-                '$token': metrics.token,
-                '$distinct_id': distinct_id
-            };
-
-            if (typeof(modifiers) === 'function') {
-                callback = modifiers;
-            }
-
-            data = merge_modifiers(data, modifiers);
-
-            if (metrics.config.debug) {
-                console.log("Sending the following data to Mixpanel (Engage):");
-                console.log(data);
-            }
-
-            metrics.send_request({ method: "GET", endpoint: "/engage", data: data }, callback);
+            const identifiers = {$distinct_id: distinct_id};
+            _union({identifiers, data, modifiers, callback, endpoint: ENGAGE_ENDPOINT, token: metrics.token, send_request: metrics.send_request, config: metrics.config});
         },
 
         /**
@@ -927,392 +654,10 @@ var create_client = function(token, config) {
             mixpanel.people.unset('bob', ['page_views', 'last_login']);
          */
         unset: function(distinct_id, prop, modifiers, callback) {
-            var $unset = [];
-
-            if (util.isArray(prop)) {
-                $unset = prop;
-            } else if (typeof(prop) === 'string') {
-                $unset = [prop];
-            } else {
-                if (metrics.config.debug) {
-                    console.error("Invalid argument passed to mixpanel.people.unset - must be a string or array");
-                    console.error("Passed: " + prop);
-                }
-                return;
-            }
-
-            var data = {
-                '$unset': $unset,
-                '$token': metrics.token,
-                '$distinct_id': distinct_id
-            };
-
-            if (typeof(modifiers) === 'function') {
-                callback = modifiers;
-            }
-
-            data = merge_modifiers(data, modifiers);
-
-            if (metrics.config.debug) {
-                console.log("Sending the following data to Mixpanel (Engage):");
-                console.log(data);
-            }
-
-            metrics.send_request({ method: "GET", endpoint: "/engage", data: data }, callback);
+            const identifiers = {$distinct_id: distinct_id};
+            _unset({identifiers, prop, modifiers, callback, endpoint: ENGAGE_ENDPOINT, token: metrics.token, send_request: metrics.send_request, config: metrics.config});
         }
     };
-
-  metrics.groups = {
-          /** groups.set_once(group_key, group_id, prop, to, modifiers, callback)
-              ---
-              The same as groups.set, but adds a property value to a group only if it has not been set before.
-
-          */
-          set_once: function(group_key, group_id, prop, to, modifiers, callback) {
-              var $set = {};
-
-              if (typeof(prop) === 'object') {
-                  if (typeof(to) === 'object') {
-                      callback = modifiers;
-                      modifiers = to;
-                  } else {
-                      callback = to;
-                  }
-                  $set = prop;
-              } else {
-                  $set[prop] = to;
-                  if (typeof(modifiers) === 'function' || !modifiers) {
-                      callback = modifiers;
-                  }
-              }
-
-              modifiers = modifiers || {};
-              modifiers.set_once = true;
-
-              this._set(group_key, group_id, $set, callback, modifiers);
-          },
-
-          /**
-              groups.set(group_key, group_id, prop, to, modifiers, callback)
-              ---
-              set properties on a group profile
-
-              usage:
-
-                  mixpanel.groups.set('company', 'Acme Inc.', 'Industry', 'Widgets');
-
-                  mixpanel.groups.set('company', 'Acme Inc.', {
-                      'Industry': 'widgets',
-                      'Plan': 'premium'
-                  });
-          */
-          set: function(group_key, group_id, prop, to, modifiers, callback) {
-              var $set = {};
-
-              if (typeof(prop) === 'object') {
-                  if (typeof(to) === 'object') {
-                      callback = modifiers;
-                      modifiers = to;
-                  } else {
-                      callback = to;
-                  }
-                  $set = prop;
-              } else {
-                  $set[prop] = to;
-                  if (typeof(modifiers) === 'function' || !modifiers) {
-                      callback = modifiers;
-                  }
-              }
-
-              this._set(group_key, group_id, $set, callback, modifiers);
-          },
-
-          // used internally by set and set_once
-          _set: function(group_key, group_id, $set, callback, options) {
-              options = options || {};
-              var set_key = (options && options.set_once) ? "$set_once" : "$set";
-
-              var data = {
-                  '$token': metrics.token,
-                  '$group_key': group_key,
-                  '$group_id': group_id
-              };
-              data[set_key] = $set;
-
-              if ('ip' in $set) {
-                  data.$ip = $set.ip;
-                  delete $set.ip;
-              }
-
-              if ($set.$ignore_time) {
-                  data.$ignore_time = $set.$ignore_time;
-                  delete $set.$ignore_time;
-              }
-
-              data = merge_modifiers(data, options);
-
-              if (metrics.config.debug) {
-                  console.log("Sending the following data to Mixpanel (Groups):");
-                  console.log(data);
-              }
-
-              metrics.send_request({ method: "GET", endpoint: "/groups", data: data }, callback);
-          },
-
-          /**
-              groups.delete_group(group_key, group_id, modifiers, callback)
-              ---
-              delete a group profile permanently
-
-              usage:
-
-                  mixpanel.groups.delete_group('company', 'Acme Inc.');
-          */
-          delete_group: function(group_key, group_id, modifiers, callback) {
-              var data = {
-                  '$delete': '',
-                  '$token': metrics.token,
-                  '$group_key': group_key,
-                  '$group_id': group_id
-              };
-
-              if (typeof(modifiers) === 'function') { callback = modifiers; }
-
-              data = merge_modifiers(data, modifiers);
-
-              if (metrics.config.debug) {
-                  console.log("Deleting group profile -", "group_key:", group_key, "group_id:", group_id);
-              }
-
-              metrics.send_request({ method: "GET", endpoint: "/groups", data: data }, callback);
-          },
-
-          /**
-             groups.remove(group_key, group_id, data, modifiers, callback)
-             ---
-             remove a value from a list-valued group profile property.
-
-             usage:
-
-                mixpanel.groups.remove('company', 'Acme Inc.', {'products': 'anvil'});
-
-                mixpanel.groups.remove('company', 'Acme Inc.', {
-                    'products': 'anvil',
-                    'customer segments': 'coyotes'
-                });
-          */
-          remove: function(group_key, group_id, data, modifiers, callback) {
-              var $remove = {};
-
-              if (typeof(data) !== 'object' || util.isArray(data)) {
-                  if (metrics.config.debug) {
-                      console.error("Invalid value passed to mixpanel.groups.remove - data must be an object with" +
-                        " scalar values");
-                  }
-                  return;
-              }
-
-              Object.keys(data).forEach(function(key) {
-                  var val = data[key];
-                  if (typeof(val) === 'string' || typeof(val) === 'number') {
-                      $remove[key] = val;
-                  } else {
-                      if (metrics.config.debug) {
-                          console.error("Invalid argument passed to mixpanel.groups.remove - values must be scalar");
-                          console.error("Passed " + key + ':', val);
-                      }
-                      return;
-                  }
-              });
-
-              if (Object.keys($remove).length === 0) {
-                  return;
-              }
-
-              data = {
-                  '$remove': $remove,
-                  '$token': metrics.token,
-                  '$group_key': group_key,
-                  '$group_id': group_id
-              };
-
-              if (typeof(modifiers) === 'function') {
-                  callback = modifiers;
-              }
-
-              data = merge_modifiers(data, modifiers);
-
-              if (metrics.config.debug) {
-                  console.log("Sending the following data to Mixpanel (Groups):");
-                  console.log(data);
-              }
-
-              metrics.send_request({ method: "GET", endpoint: "/groups", data: data }, callback);
-          },
-
-          /**
-           groups.union(group_key, group_id, data, modifiers, callback)
-           ---
-           merge value(s) into a list-valued group profile property.
-
-           usage:
-
-              mixpanel.groups.union('company', 'Acme Inc.', {'products': 'anvil'});
-
-              mixpanel.groups.union('company', 'Acme Inc.', {'products': ['anvil'], 'customer segments': ['coyotes']});
-           */
-          union: function(group_key, group_id, data, modifiers, callback) {
-              var $union = {};
-
-              if (typeof(data) !== 'object' || util.isArray(data)) {
-                  if (metrics.config.debug) {
-                      console.error("Invalid value passed to mixpanel.groups.union - data must be an object with" +
-                        " scalar" +
-                        " or array values");
-                  }
-                  return;
-              }
-
-              Object.keys(data).forEach(function(key) {
-                  var val = data[key];
-                  if (util.isArray(val)) {
-                      var merge_values = val.filter(function(v) {
-                          return typeof(v) === 'string' || typeof(v) === 'number';
-                      });
-                      if (merge_values.length > 0) {
-                          $union[key] = merge_values;
-                      }
-                  } else if (typeof(val) === 'string' || typeof(val) === 'number') {
-                      $union[key] = [val];
-                  } else {
-                      if (metrics.config.debug) {
-                          console.error("Invalid argument passed to mixpanel.groups.union - values must be a scalar" +
-                            " value" +
-                            " or array");
-                          console.error("Passed " + key + ':', val);
-                      }
-                      return;
-                  }
-              });
-
-              if (Object.keys($union).length === 0) {
-                  return;
-              }
-
-              data = {
-                  '$union': $union,
-                  '$token': metrics.token,
-                  '$group_key': group_key,
-                  '$group_id': group_id
-              };
-
-              if (typeof(modifiers) === 'function') {
-                  callback = modifiers;
-              }
-
-              data = merge_modifiers(data, modifiers);
-
-              if (metrics.config.debug) {
-                  console.log("Sending the following data to Mixpanel (Groups):");
-                  console.log(data);
-              }
-
-              metrics.send_request({ method: "GET", endpoint: "/groups", data: data }, callback);
-          },
-
-          /**
-           groups.unset(group_key, group_id, prop, modifiers, callback)
-           ---
-           delete a property on a group profile
-
-           usage:
-
-              mixpanel.groups.unset('company', 'Acme Inc.', 'products');
-
-              mixpanel.groups.unset('company', 'Acme Inc.', ['products', 'customer segments']);
-           */
-          unset: function(group_key, group_id, prop, modifiers, callback) {
-              var $unset = [];
-
-              if (util.isArray(prop)) {
-                  $unset = prop;
-              } else if (typeof(prop) === 'string') {
-                  $unset = [prop];
-              } else {
-                  if (metrics.config.debug) {
-                      console.error("Invalid argument passed to mixpanel.groups.unset - must be a string or array");
-                      console.error("Passed: " + prop);
-                  }
-                  return;
-              }
-
-              var data = {
-                  '$unset': $unset,
-                  '$token': metrics.token,
-                  '$group_key': group_key,
-                  '$group_id': group_id
-              };
-
-              if (typeof(modifiers) === 'function') {
-                  callback = modifiers;
-              }
-
-              data = merge_modifiers(data, modifiers);
-
-              if (metrics.config.debug) {
-                  console.log("Sending the following data to Mixpanel (Groups):");
-                  console.log(data);
-              }
-
-              metrics.send_request({ method: "GET", endpoint: "/groups", data: data }, callback);
-          }
-      };
-
-    var merge_modifiers = function(data, modifiers) {
-        if (modifiers) {
-            if (modifiers.$ignore_alias) {
-                data.$ignore_alias = modifiers.$ignore_alias;
-            }
-            if (modifiers.$ignore_time) {
-                data.$ignore_time = modifiers.$ignore_time;
-            }
-            if (modifiers.hasOwnProperty("$ip")) {
-                data.$ip = modifiers.$ip;
-            }
-            if (modifiers.hasOwnProperty("$time")) {
-                data.$time = ensure_timestamp(modifiers.$time);
-            }
-        }
-        return data;
-    };
-
-    /**
-        set_config(config)
-        ---
-        Modifies the mixpanel config
-
-        config:object       an object with properties to override in the
-                            mixpanel client config
-    */
-    metrics.set_config = function(config) {
-        for (var c in config) {
-            if (config.hasOwnProperty(c)) {
-                if (c == "host") { // Split host, into host and port.
-                    metrics.config.host = config[c].split(':')[0];
-                    var port = config[c].split(':')[1];
-                    if (port) {
-                        metrics.config.port = Number(port);
-                    }
-                } else {
-                    metrics.config[c] = config[c];
-                }
-            }
-        }
-    };
-
-    if (config) {
-        metrics.set_config(config);
-    }
 
     return metrics;
 };
